requirements.txt
#install all requirements with "pip install -r requirements.txt"

openai
langchain
pypdf
tiktoken
python-dotenv
chromadb
bs4